Add name to container, add guard to remove.sh

+++ remove.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-container_id=$(docker ps -aqf "ancestor=dev-docker")
-docker stop $container_id
-docker rm $container_id
+container_id=$(docker ps -aqf "name=dev-docker")
+if [[ ! -z $container_id ]]; then
+  docker stop $container_id
+  docker rm $container_id
+fi
